moving Console service inside the bootstrap
